rename sqlite_test.go to adapter_test.go
